Add experimental TravisCI support for python v3.4
Eventually we want to be able to run roundup with python 3, so we may as
well run the tests under python 3 to get an idea of how close support
is. The python 3 builds are allowed to fail, so any issues running the
tests under python 3 won't report the python 2 tests as failing.

Because there will be enough issues with the core code we aren't
bothering to run the tests with dependencies. Once we have debugged all
the core issues we can look to add the dependencies back in.