Add a open() method
Opening open the database and starting a transaction in the __init__
method seems broken, but that is how the existing roundup backends
currently work. We will look at seeing if we can change the roundup
internals later, but for the time being create a open() method and use
it within the __init__() method.