Remove the tsearch2 backend
The documentation within the backend labels it as being experimental,
and that it should be used. The roundup.backend.have_backend() returns
that it does not exist and is labeled as "currently not working". The
PostgreSQL website also seems to indicate that it has been deprecated
since v8.3 when text searching was integrated into the core[1].

After considering all this it seems like the best option is to just
remove the backend.

[1] http://www.postgresql.org/docs/9.4/static/tsearch2.html