Add mailgw.RoundupMessage class
The existing mailgw.Message class is based on mimetools.Message, but
mimetools has been deprecated since Python v2.3. The intention is to
replace mailgw.Message with mailgw.RoundupMessage which is based on
email.message.Message and should provide better compatibility with
any moves to provide Roundup support for Python v3.