Cleanup test/test_mailgw.py test email messages
A number of obvious errors in the test email messages can be cleaned up
to look more like valid emails instead of relying on the behaviour of
mimetools.Message to interpret them correctly. The email.message.Message
class is also know to not parse multi-line header strings if there is no
leading whitespace on the second and subsequent lines of a header
string, so these need to be fixed before we can migrate away from
the mimetools module to the email module.