Make @template support two alternate templates for error and ok cases.
Setting @template=oktmpl|errortmpl in a form will display the next
page using the oktmpl if the change did not cause an error. If
submitting the form caused an error (raised by an auditor or something
else), the user is displayed a page using the errortmpl. Docs in
customizing.tmpl. Look for modal edit.