Three sets of changes:
  1) Make sure that a user doesn't create a query with the same name
     as an existing query that the user owns.

  2) When submitting a new named query, display the name of the query
     on the index page.

  3) Allow optional arguments to indexargs_url by setting their value
     to None. This will show the argument only if there is a valid
     value.

To match these, changed the template for all search templates so if an
error is thrown due to #1 the user stays on the search page so they
can fix the issue.

Note that I did not add automated tests for these because I couldn't
find existing tests for these code paths that I could adapt. I don't
understand how the existing Action tests work and there is no doc for
them.